Refactored Edit Command
