+++ src/main/java/seedu/address/logic/commands/EditCommand.java
@@ -33,6 +33,9 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.CollectionUtil;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.descriptors.EditAppointmentDescriptor;
+import seedu.address.logic.descriptors.EditPersonDescriptor;
+import seedu.address.logic.descriptors.EditPetPatientDescriptor;
 import seedu.address.model.appointment.Appointment;
 import seedu.address.model.appointment.Remark;
 import seedu.address.model.appointment.exceptions.AppointmentNotFoundException;
@@ -465,353 +468,5 @@ public class EditCommand extends UndoableCommand {
         return index.equals(e.index)
                 && editPersonDescriptor.equals(e.editPersonDescriptor)
                 && Objects.equals(personToEdit, e.personToEdit);
-    }
-
-    /**
-     * Stores the details to edit the person with. Each non-empty field value will replace the
-     * corresponding field value of the person.
-     */
-    public static class EditPersonDescriptor {
-        private Name name;
-        private Phone phone;
-        private Email email;
-        private Address address;
-        private Nric nric;
-        private Set<Tag> tags;
-
-        public EditPersonDescriptor() {}
-
-        /**
-         * Copy constructor.
-         * A defensive copy of {@code tags} is used internally.
-         */
-        public EditPersonDescriptor(EditPersonDescriptor toCopy) {
-            setName(toCopy.name);
-            setPhone(toCopy.phone);
-            setEmail(toCopy.email);
-            setAddress(toCopy.address);
-            setNric(toCopy.nric);
-            setTags(toCopy.tags);
-        }
-
-        /**
-         * Returns true if at least one field is edited.
-         */
-        public boolean isAnyFieldEdited() {
-            return CollectionUtil.isAnyNonNull(this.name, this.phone, this.email,
-                    this.address, this.nric, this.tags);
-        }
-
-        public void setName(Name name) {
-            this.name = name;
-        }
-
-        public Optional<Name> getName() {
-            return Optional.ofNullable(name);
-        }
-
-        public void setPhone(Phone phone) {
-            this.phone = phone;
-        }
-
-        public Optional<Phone> getPhone() {
-            return Optional.ofNullable(phone);
-        }
-
-        public void setEmail(Email email) {
-            this.email = email;
-        }
-
-        public Optional<Email> getEmail() {
-            return Optional.ofNullable(email);
-        }
-
-        public void setAddress(Address address) {
-            this.address = address;
-        }
-
-        public Optional<Address> getAddress() {
-            return Optional.ofNullable(address);
-        }
-
-        public void setNric(Nric nric) {
-            this.nric = nric;
-        }
-
-        public Optional<Nric> getNric() {
-            return Optional.ofNullable(nric);
-        }
-
-        /**
-         * Sets {@code tags} to this object's {@code tags}.
-         * A defensive copy of {@code tags} is used internally.
-         */
-        public void setTags(Set<Tag> tags) {
-            this.tags = (tags != null) ? new HashSet<>(tags) : null;
-        }
-
-        /**
-         * Returns an unmodifiable tag set, which throws {@code UnsupportedOperationException}
-         * if modification is attempted.
-         * Returns {@code Optional#empty()} if {@code tags} is null.
-         */
-        public Optional<Set<Tag>> getTags() {
-            return (tags != null) ? Optional.of(Collections.unmodifiableSet(tags)) : Optional.empty();
-        }
-
-        @Override
-        public boolean equals(Object other) {
-            // short circuit if same object
-            if (other == this) {
-                return true;
-            }
-
-            // instanceof handles nulls
-            if (!(other instanceof EditPersonDescriptor)) {
-                return false;
-            }
-
-            // state check
-            EditPersonDescriptor e = (EditPersonDescriptor) other;
-
-            return getName().equals(e.getName())
-                    && getPhone().equals(e.getPhone())
-                    && getEmail().equals(e.getEmail())
-                    && getAddress().equals(e.getAddress())
-                    && getNric().equals(e.getNric())
-                    && getTags().equals(e.getTags());
-        }
-    }
-
-    /**
-     * Stores the details to edit the pet patient with. Each non-empty field value will replace the
-     * corresponding field value of the pet patient.
-     */
-    public static class EditPetPatientDescriptor {
-        private PetPatientName name;
-        private Species species;
-        private Breed breed;
-        private Colour colour;
-        private BloodType bloodType;
-        private Nric ownerNric;
-        private Set<Tag> tags;
-
-        public EditPetPatientDescriptor() {}
-
-        /**
-         * Copy constructor.
-         * A defensive copy of {@code tags} is used internally.
-         */
-        public EditPetPatientDescriptor(EditPetPatientDescriptor toCopy) {
-            setName(toCopy.name);
-            setSpecies(toCopy.species);
-            setBreed(toCopy.breed);
-            setColour(toCopy.colour);
-            setBloodType(toCopy.bloodType);
-            setOwnerNric(toCopy.ownerNric);
-            setTags(toCopy.tags);
-        }
-
-        /**
-         * Returns true if at least one field is edited.
-         */
-        public boolean isAnyFieldEdited() {
-            return CollectionUtil.isAnyNonNull(this.name, this.species, this.breed,
-                    this.colour, this.bloodType, this.ownerNric, this.tags);
-        }
-
-        public void setName(PetPatientName name) {
-            this.name = name;
-        }
-
-        public Optional<PetPatientName> getName() {
-            return Optional.ofNullable(name);
-        }
-
-        public void setSpecies(Species species) {
-            this.species = species;
-        }
-
-        public Optional<Species> getSpecies() {
-            return Optional.ofNullable(species);
-        }
-
-        public void setBreed(Breed breed) {
-            this.breed = breed;
-        }
-
-        public Optional<Breed> getBreed() {
-            return Optional.ofNullable(breed);
-        }
-
-        public void setColour(Colour colour) {
-            this.colour = colour;
-        }
-
-        public Optional<Colour> getColour() {
-            return Optional.ofNullable(colour);
-        }
-
-        public void setBloodType(BloodType bloodType) {
-            this.bloodType = bloodType;
-        }
-
-        public Optional<BloodType> getBloodType() {
-            return Optional.ofNullable(bloodType);
-        }
-
-        public void setOwnerNric(Nric nric) {
-            this.ownerNric = nric;
-        }
-
-        public Optional<Nric> getOwnerNric() {
-            return Optional.ofNullable(ownerNric);
-        }
-
-        /**
-         * Sets {@code tags} to this object's {@code tags}.
-         * A defensive copy of {@code tags} is used internally.
-         */
-        public void setTags(Set<Tag> tags) {
-            this.tags = (tags != null) ? new HashSet<>(tags) : null;
-        }
-
-        /**
-         * Returns an unmodifiable tag set, which throws {@code UnsupportedOperationException}
-         * if modification is attempted.
-         * Returns {@code Optional#empty()} if {@code tags} is null.
-         */
-        public Optional<Set<Tag>> getTags() {
-            return (tags != null) ? Optional.of(Collections.unmodifiableSet(tags)) : Optional.empty();
-        }
-
-        @Override
-        public boolean equals(Object other) {
-            // short circuit if same object
-            if (other == this) {
-                return true;
-            }
-
-            // instanceof handles nulls
-            if (!(other instanceof EditPetPatientDescriptor)) {
-                return false;
-            }
-
-            // state check
-            EditPetPatientDescriptor e = (EditPetPatientDescriptor) other;
-
-            return getName().equals(e.getName())
-                    && getSpecies().equals(e.getSpecies())
-                    && getBreed().equals(e.getBreed())
-                    && getColour().equals(e.getColour())
-                    && getBloodType().equals(e.getBloodType())
-                    && getOwnerNric().equals(e.getOwnerNric())
-                    && getTags().equals(e.getTags());
-        }
-    }
-
-    /**
-     * Stores the details to edit the appointment with. Each non-empty field value will replace the
-     * corresponding field value of the appointment.
-     */
-    public static class EditAppointmentDescriptor {
-        private Nric ownerNric;
-        private PetPatientName petPatientName;
-        private Remark remark;
-        private LocalDateTime localDateTime;
-        private Set<Tag> tags;
-
-        public EditAppointmentDescriptor() {}
-
-        /**
-         * Copy constructor.
-         * A defensive copy of {@code tags} is used internally.
-         */
-        public EditAppointmentDescriptor(EditAppointmentDescriptor toCopy) {
-            setOwnerNric(toCopy.ownerNric);
-            setPetPatientName(toCopy.petPatientName);
-            setRemark(toCopy.remark);
-            setLocalDateTime(toCopy.localDateTime);
-            setTags(toCopy.tags);
-        }
-
-        /**
-         * Returns true if at least one field is edited.
-         */
-        public boolean isAnyFieldEdited() {
-            return CollectionUtil.isAnyNonNull(this.ownerNric, this.petPatientName, this.remark,
-                    this.localDateTime, this.tags);
-        }
-
-        public void setOwnerNric(Nric ownerNric) {
-            this.ownerNric = ownerNric;
-        }
-
-        public Optional<Nric> getOwnerNric() {
-            return Optional.ofNullable(ownerNric);
-        }
-
-        public void setPetPatientName(PetPatientName petPatientName) {
-            this.petPatientName = petPatientName;
-        }
-
-        public Optional<PetPatientName> getPetPatientName() {
-            return Optional.ofNullable(petPatientName);
-        }
-
-        public void setRemark(Remark remark) {
-            this.remark = remark;
-        }
-
-        public Optional<Remark> getRemark() {
-            return Optional.ofNullable(remark);
-        }
-
-        public void setLocalDateTime(LocalDateTime localDateTime) {
-            this.localDateTime = localDateTime;
-        }
-
-        public Optional<LocalDateTime> getLocalDateTime() {
-            return Optional.ofNullable(localDateTime);
-        }
-
-        /**
-         * Sets {@code tags} to this object's {@code tags}.
-         * A defensive copy of {@code tags} is used internally.
-         */
-        public void setTags(Set<Tag> tags) {
-            this.tags = (tags != null) ? new HashSet<>(tags) : null;
-        }
-
-        /**
-         * Returns an unmodifiable tag set, which throws {@code UnsupportedOperationException}
-         * if modification is attempted.
-         * Returns {@code Optional#empty()} if {@code tags} is null.
-         */
-        public Optional<Set<Tag>> getTags() {
-            return (tags != null) ? Optional.of(Collections.unmodifiableSet(tags)) : Optional.empty();
-        }
-
-        @Override
-        public boolean equals(Object other) {
-            // short circuit if same object
-            if (other == this) {
-                return true;
-            }
-
-            // instanceof handles nulls
-            if (!(other instanceof EditAppointmentDescriptor)) {
-                return false;
-            }
-
-            // state check
-            EditAppointmentDescriptor e = (EditAppointmentDescriptor) other;
-
-            return getOwnerNric().equals(e.getOwnerNric())
-                    && getPetPatientName().equals(e.getPetPatientName())
-                    && getRemark().equals(e.getRemark())
-                    && getLocalDateTime().equals(e.getLocalDateTime())
-                    && getTags().equals(e.getTags());
-        }
     }
 }
