+++ src/main/java/seedu/address/logic/descriptors/EditPersonDescriptor.java
@@ -1,0 +1,131 @@
+package seedu.address.logic.descriptors;
+
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Optional;
+import java.util.Set;
+
+import seedu.address.commons.util.CollectionUtil;
+import seedu.address.model.person.Address;
+import seedu.address.model.person.Email;
+import seedu.address.model.person.Name;
+import seedu.address.model.person.Nric;
+import seedu.address.model.person.Phone;
+import seedu.address.model.tag.Tag;
+
+//@@author chialejing-reused
+/**
+ * Stores the details to edit the person with. Each non-empty field value will replace the
+ * corresponding field value of the person.
+ */
+public class EditPersonDescriptor {
+    private Name name;
+    private Phone phone;
+    private Email email;
+    private Address address;
+    private Nric nric;
+    private Set<Tag> tags;
+
+    public EditPersonDescriptor() {}
+
+    /**
+     * Copy constructor.
+     * A defensive copy of {@code tags} is used internally.
+     */
+    public EditPersonDescriptor(EditPersonDescriptor toCopy) {
+        setName(toCopy.name);
+        setPhone(toCopy.phone);
+        setEmail(toCopy.email);
+        setAddress(toCopy.address);
+        setNric(toCopy.nric);
+        setTags(toCopy.tags);
+    }
+
+    /**
+     * Returns true if at least one field is edited.
+     */
+    public boolean isAnyFieldEdited() {
+        return CollectionUtil.isAnyNonNull(this.name, this.phone, this.email,
+                this.address, this.nric, this.tags);
+    }
+
+    public void setName(Name name) {
+        this.name = name;
+    }
+
+    public Optional<Name> getName() {
+        return Optional.ofNullable(name);
+    }
+
+    public void setPhone(Phone phone) {
+        this.phone = phone;
+    }
+
+    public Optional<Phone> getPhone() {
+        return Optional.ofNullable(phone);
+    }
+
+    public void setEmail(Email email) {
+        this.email = email;
+    }
+
+    public Optional<Email> getEmail() {
+        return Optional.ofNullable(email);
+    }
+
+    public void setAddress(Address address) {
+        this.address = address;
+    }
+
+    public Optional<Address> getAddress() {
+        return Optional.ofNullable(address);
+    }
+
+    public void setNric(Nric nric) {
+        this.nric = nric;
+    }
+
+    public Optional<Nric> getNric() {
+        return Optional.ofNullable(nric);
+    }
+
+    /**
+     * Sets {@code tags} to this object's {@code tags}.
+     * A defensive copy of {@code tags} is used internally.
+     */
+    public void setTags(Set<Tag> tags) {
+        this.tags = (tags != null) ? new HashSet<>(tags) : null;
+    }
+
+    /**
+     * Returns an unmodifiable tag set, which throws {@code UnsupportedOperationException}
+     * if modification is attempted.
+     * Returns {@code Optional#empty()} if {@code tags} is null.
+     */
+    public Optional<Set<Tag>> getTags() {
+        return (tags != null) ? Optional.of(Collections.unmodifiableSet(tags)) : Optional.empty();
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        // short circuit if same object
+        if (other == this) {
+            return true;
+        }
+
+        // instanceof handles nulls
+        if (!(other instanceof EditPersonDescriptor)) {
+            return false;
+        }
+
+        // state check
+        EditPersonDescriptor e = (EditPersonDescriptor) other;
+
+        return getName().equals(e.getName())
+                && getPhone().equals(e.getPhone())
+                && getEmail().equals(e.getEmail())
+                && getAddress().equals(e.getAddress())
+                && getNric().equals(e.getNric())
+                && getTags().equals(e.getTags());
+    }
+}
